+++ R/sampling.R
@@ -83,7 +83,7 @@
 #'   require(ggplot2, quietly = TRUE) &&
 #'   bru_safe_sp() &&
 #'   require("sp")) {
-#'   data("gorillas", package = "inlabru")
+#'   gorillas <- gorillas_sp()
 #'   pts <- sample.lgcp(gorillas$mesh,
 #'     loglambda = 1.5,
 #'     samplers = gorillas$boundary