+++ tests/testthat/test-lgcp_2d_plotsample.R
@@ -8,7 +8,7 @@ test_that("2D LGCP fitting and prediction: Plot sampling", {
       int.strategy = "eb"
     )
   )
-  data(gorillas, package = "inlabru", envir = environment())
+  gorillas <- gorillas_sp()
 
   matern <- INLA::inla.spde2.pcmatern(gorillas$mesh,
     prior.sigma = c(0.1, 0.01),