+++ NEWS.md
@@ -35,6 +35,8 @@
 * Remove `seals_sp` data set due to excessive size (version `2.11.1.9010`)
 * Replace the `mexdolphin` dataset with a function `mexdolphin_sp()`
   to avoid `sp` data objects in the package (version `2.11.1.9015`)
+* Replace the `gorillas` dataset with a function `gorillas_sp()`
+  to avoid `sp` data objects in the package (version `2.11.1.9016`)
 
 # inlabru 2.11.1